+++ src/document/url.ml
@@ -121,6 +121,10 @@ module Path = struct
 
   let pp_kind fmt kind = Format.fprintf fmt "%s" (string_of_kind kind)
 
+  let pp_disambiguating_prefix fmt = function
+    | `Module | `Page | `LeafPage | `File | `SourcePage -> ()
+    | kind -> Format.fprintf fmt "%s-" (string_of_kind kind)
+
   type t = { kind : kind; parent : t option; name : string }
 
   let mk ?parent kind name = { kind; parent; name }