+++ src/html/link.ml
@@ -7,9 +7,7 @@ module Path = struct
   let for_printing url = List.map snd @@ Url.Path.to_list url
 
   let segment_to_string (kind, name) =
-    match kind with
-    | `Module | `Page | `File | `SourcePage -> name
-    | _ -> Format.asprintf "%a-%s" Url.Path.pp_kind kind name
+    Format.asprintf "%a%s" Url.Path.pp_disambiguating_prefix kind name
 
   let is_leaf_page url = url.Url.Path.kind = `LeafPage